CLOUDP-239577: [AtlasCLI] Move deployment connect opts into connect command (#3109)
